init logic in operation enabled is changed

+++ CSVMotorController/src/CSVMotorController.h
@@ -22,6 +22,6 @@ public:
     void setTrajectory(float tick);
     float getVelTick();
     float getPosTick();
-    void logging(int tick, uint32_t actualVel, uint32_t actualPos);
+    void logging(float tick, uint32_t actualVel, uint32_t actualPos);
     void saveData(const string &position_filename, const string &velocity_filename);    
 };
